refactor: extract workflow engine to src/workflow module
- Move generic workflow engine files from src/core/workflow/ to src/workflow/
- Move domain-specific steps from src/core/workflow/steps/ to src/core/steps/
- Update all imports throughout codebase
- Clean separation between reusable workflow engine and domain logic
