+++ sgl-model-gateway/src/core/job_queue.rs
@@ -16,16 +16,14 @@ use tracing::{debug, error, info, warn};
 use crate::{
     app_context::AppContext,
     config::{RouterConfig, RoutingMode},
-    core::workflow::{
-        steps::{
-            McpServerConfigRequest, WasmModuleConfigRequest, WasmModuleRemovalRequest,
-            WorkerRemovalRequest,
-        },
-        WorkflowContext, WorkflowEngine, WorkflowId, WorkflowInstanceId, WorkflowStatus,
+    core::steps::{
+        McpServerConfigRequest, WasmModuleConfigRequest, WasmModuleRemovalRequest,
+        WorkerRemovalRequest,
     },
     mcp::McpConfig,
     observability::metrics::RouterMetrics,
     protocols::worker_spec::{JobStatus, WorkerConfigRequest},
+    workflow::{WorkflowContext, WorkflowEngine, WorkflowId, WorkflowInstanceId, WorkflowStatus},
 };
 
 /// Job types for control plane operations
