+++ sgl-model-gateway/src/lib.rs
@@ -17,3 +17,4 @@ pub mod tokenizer;
 pub mod tool_parser;
 pub mod version;
 pub mod wasm;
+pub mod workflow;
