+++ sgl-model-gateway/src/workflow/mod.rs
@@ -5,7 +5,6 @@ mod engine;
 mod event;
 mod executor;
 mod state;
-pub mod steps;
 pub mod types;
 
 // Re-export main types
@@ -14,9 +13,4 @@ pub use engine::WorkflowEngine;
 pub use event::{EventBus, EventSubscriber, LoggingSubscriber, WorkflowEvent};
 pub use executor::{FunctionStep, StepExecutor};
 pub use state::WorkflowStateStore;
-pub use steps::{
-    create_external_worker_registration_workflow, create_mcp_registration_workflow,
-    create_wasm_module_registration_workflow, create_wasm_module_removal_workflow,
-    create_worker_registration_workflow, create_worker_removal_workflow,
-};
 pub use types::*;
